Do not kill more processes than we have to
The process that is known to have terminated need not be killed again.
We might still end up killing processes that have already terminated
if they finished very shortly after the first one (this is rather
likely to happen because the workload on each process should be
approximately the same).